Added some docstrings to the life cycle test
