+++ tests/test_misc.py
@@ -5,11 +5,13 @@ Description: Implements test cases
 Author(s): Pooya Hekmati, sh2235@cornell.edu
 """
 
-import mock
-from mock.mock import MagicMock
-from pytest_mock.plugin import MockerFixture
 import pytest
+from mock.mock import MagicMock
+from pytest import approx, raises
+from pytest_mock.plugin import MockerFixture
+
 from RUFAS.simulation_engine import SimulationEngine
+from RUFAS.util import Utility
 
 
 def test_annual_reset():
@@ -56,7 +58,7 @@ def test_is_leap_year():
 def patch_simulation_engine(mocker: MockerFixture) -> SimulationEngine:
     """Returns a mocked SimulationEngine"""
     mocker.patch(
-        'RUFAS.simulation_engine.SimulationEngine._initialize_simulation')
+            'RUFAS.simulation_engine.SimulationEngine._initialize_simulation')
 
     sim_eng = SimulationEngine('dummy_path')
     sim_eng.config = MagicMock()
@@ -71,20 +73,20 @@ def patch_simulation_engine(mocker: MockerFixture) -> SimulationEngine:
 def test_init_simulation_engine(patch_simulation_engine: SimulationEngine) -> None:
     """Unit test for function __init__ in file RUFAS/simulation_engine.py"""
     patch_simulation_engine._initialize_simulation.assert_called_once_with(
-        'dummy_path')
+            'dummy_path')
 
 
 def test_simulate(patch_simulation_engine: SimulationEngine, mocker: MockerFixture) -> None:
     """Unit test for function simulate in file RUFAS/simulation_engine.py"""
     mocker.patch(
-        'RUFAS.simulation_engine.SimulationEngine._run_simulation_main_loop')
+            'RUFAS.simulation_engine.SimulationEngine._run_simulation_main_loop')
     mocker.patch(
-        'RUFAS.simulation_engine.SimulationEngine._show_final_messages')
+            'RUFAS.simulation_engine.SimulationEngine._show_final_messages')
     sim_eng = patch_simulation_engine
     sim_eng.simulate()
     sim_eng._run_simulation_main_loop.assert_called_once()
     sim_eng.output.finalize.assert_called_once_with(
-        sim_eng.state, sim_eng.weather, sim_eng.time)
+            sim_eng.state, sim_eng.weather, sim_eng.time)
     sim_eng._show_final_messages.assert_called_once()
 
 
@@ -181,3 +183,104 @@ def test_organize_results():
 def test_LP_print():
     """Unit test for function LP_print in file util.py"""
     pass
+
+
+def test_calc_average() -> None:
+    """Unit test for function calc_average in file util.py"""
+    # Normal case
+    result = Utility.calc_average(num_values=9, cur_avg=5, new_value=6)
+    actual_new_num_values, actual_new_avg = result
+    assert actual_new_num_values == 10
+    assert actual_new_avg == approx(5.1)  # (9 * 5 + 6) / 10
+
+    # Given a count of 0 and an average value of 0.0,
+    # the function should return whatever the new value is.
+    result = Utility.calc_average(num_values=0, cur_avg=0.0, new_value=6.0)
+    actual_new_num_values, actual_new_avg = result
+    assert actual_new_num_values == 1
+    assert actual_new_avg == approx(6.0)
+
+
+def test_remove_items_from_list_by_indices() -> None:
+    """Unit test for function remove_items_from_list_by_indices in file util.py"""
+    # Given an empty list and an empty list of removal indices,
+    # the function should do nothing.
+    arr = []
+    del_idx = []
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert len(arr) == 0
+
+    # Given a non-empty list and an empty list of removal indices,
+    # the function should do nothing.
+    arr = [0, 1, 2]
+    del_idx = []
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert arr == [0, 1, 2]
+
+    # Given a list of size 1 and the removal index of 0,
+    # the function should return an empty list.
+    arr = [0]
+    del_idx = [0]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert len(arr) == 0
+
+    # Given a list of size 2 and one valid removal index,
+    # the function should return a correct list of size 1.
+    arr = [10, 20]
+    del_idx = [0]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert arr == [20]
+
+    arr = [10, 20]
+    del_idx = [1]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert arr == [10]
+
+    # Given a list of size 3 and a list of 2 removal indices,
+    # the function should return a correct list of size 1.
+    arr = [10, 20, 30]
+    del_idx = [0, 1]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert arr == [30]
+
+    arr = [10, 20, 30]
+    del_idx = [1, 2]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert arr == [10]
+
+    arr = [10, 20, 30]
+    del_idx = [0, 2]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
+    assert arr == [20]
+
+    # Given an empty list and a non-empty list of removal indices,
+    # the function should raise IndexError.
+    arr = []
+    del_idx = [0]
+    with raises(IndexError):
+        Utility.remove_items_from_list_by_indices(arr, del_idx)
+
+
+def test_percent_calculator() -> None:
+    """Unit test for function percent_calculator in file util.py"""
+    # Normal case
+    # Given any random non-zero denominator,
+    # the function should return correct percentages.
+    pc = Utility.percent_calculator(denominator=20)
+    assert pc(0) == approx(0.0)
+    assert pc(20) == approx(100.0)
+    assert pc(8) == approx(40.0)  # e.g., 8/20 = 40%
+    assert pc(-8) == approx(-40.0)
+    assert pc(24) == approx(120.0)
+
+    # Given a denominator of 100,
+    # the function should return the numerator as percentage.
+    pc = Utility.percent_calculator(denominator=100)
+    assert pc(0.0) == approx(0.0)
+    assert pc(12.3) == approx(12.3)
+    assert pc(100.0) == approx(100.0)
+
+    # Given a 0 denominator, the function should raise a ZeroDivisionError.
+    pc = Utility.percent_calculator(denominator=0)
+    with raises(ZeroDivisionError):
+        pc(1.0)
